refactor: centralize type system and optimize import structure
- Consolidate scattered type definitions into /types/ directory
- Remove redundant type files from feature directories
- Update all import paths to use centralized types
- Improve type organization for better maintainability
- Add error types to centralized location
