+++ app/(admin)/admin/jobs/components/JobsHeader.tsx
@@ -2,7 +2,7 @@ import { Briefcase } from 'lucide-react'
 
 import { AdminHeader } from '@/components/ui/admin-header'
 
-import type { JobCounts } from '../types'
+import type { JobCounts } from '@/types'
 
 interface JobsHeaderProps {
   counts: JobCounts | null
