+++ app/(admin)/admin/jobs/components/JobsTabs.tsx
@@ -1,6 +1,6 @@
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 
-import type { JobItem, JobCounts, JobStatus } from '../types'
+import type { AdminJobItem as JobItem, JobCounts, JobStatus } from '@/types'
 
 import { JobsTable } from './JobsTable'
