+++ app/api/admin/jobs/route.ts
@@ -3,7 +3,7 @@ import { createClient } from '@supabase/supabase-js'
 
 import { createServerSupabaseClient } from '@/lib/supabase/server'
 
-import type { Database } from '@/lib/types/database'
+import type { Database } from '@/types'
 
 // Middleware to check if user is admin
 async function checkAdminAuth() {
