+++ app/api/applications/route.ts
@@ -5,7 +5,7 @@ import { EmailService } from '@/lib/email/service'
 import { createRouteSupabaseClient } from '@/lib/supabase/server'
 import { applicationSchema } from '@/lib/validations/application'
 
-import type { Database } from '@/lib/types/database'
+import type { Database } from '@/types'
 
 // Remove legacy application detection since we're using the new schema
