+++ lib/database/connection-pool.ts
@@ -1,5 +1,5 @@
 import type { SupabaseClient } from '@supabase/supabase-js'
-import type { Database } from '@/lib/types/database'
+import type { Database } from '@/types'
 
 // Connection pool interface
 interface ConnectionPool {
