+++ lib/supabase/server.ts
@@ -1,7 +1,7 @@
 import { createServerClient } from '@supabase/ssr'
 import { createClient } from '@supabase/supabase-js'
 import { cookies } from 'next/headers'
-import type { Database } from '@/types/database'
+import type { Database } from '@/types'
 
 // Validate environment variables on module load
 const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL!
