--- lib/types/index.ts
@@ -1,14 +1,0 @@
-// Database types
-export type { Database } from './database'
-
-// Shared admin types
-export type {
-  StatusType,
-  StatusWithAll,
-  Counts,
-  ApiError,
-  ApiResponse,
-  FormErrors,
-  AdminItem,
-  PaginationOptions
-} from './shared'
