+++ types/api/errors.ts
@@ -1,0 +1,31 @@
+// Error handling utilities for API operations
+
+export function handleJobsError(error: unknown): string {
+  if (error instanceof Error) {
+    return error.message
+  }
+  if (typeof error === 'string') {
+    return error
+  }
+  return 'An unexpected error occurred while processing jobs'
+}
+
+export function handleApplicationsError(error: unknown): string {
+  if (error instanceof Error) {
+    return error.message
+  }
+  if (typeof error === 'string') {
+    return error
+  }
+  return 'An unexpected error occurred while processing applications'
+}
+
+export function handleApiError(error: unknown): string {
+  if (error instanceof Error) {
+    return error.message
+  }
+  if (typeof error === 'string') {
+    return error
+  }
+  return 'An unexpected error occurred'
+}
